Refactoring: move spiralPointGenerator to CircularCloudLayouter
